Move api changeset upload-download tests
